requirements.txt
flask==3.0.0
openai>=1.50.0
python-dotenv==1.0.0
werkzeug==3.0.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
requests>=2.32.0
gunicorn==21.2.0 